Move errors to separate file, add extension with error descriptions.
